pixel cycle to tag direction

+++ vrc_sandbox/sandbox.py
@@ -253,7 +253,7 @@ class Sandbox():
             tagid = tag["id"]
 
             heading = tag["heading"] 
-            which_pixel = (int)(heading/360)*32
+            which_pixel = (int)(((float)(heading/360.0))*32.0)
                 #{"target_pixel":16,"delay_ms":250}
                 #vrc/pcc/set_pixel_cycle
             datamsgpixl = {"target_pixel": which_pixel, "delay_ms": 25}
